+++ packages/terra-clinical-item-display/src/terra-dev-site/doc/clinical-item-display/clinicalItemDisplay.1.doc.mdx
@@ -10,7 +10,7 @@ import Icon from '../example/Icon?dev-site-example';
 import IconText from '../example/IconText?dev-site-example';
 import DefaultComment from '../example/DefaultComment?dev-site-example';
 import TextStyleMeaning from '../example/TextStyleMeaning?dev-site-example';
-import TruncatedTextPopup from '../example/TruncatedTextPopup?dev-site-example';
+import TruncatedTextExample from '../example/TruncatedText?dev-site-example';
 
 <Badge />
 
@@ -82,7 +82,7 @@ experience for all users.
 <IconText title="Item Display: Icon and Text" />
 <DefaultComment title="Comment Item Display" />
 <TextStyleMeaning title="Item Display: Text Style Meaning for Assistive Technology" />
-<TruncatedTextPopup title="Item Display: Truncated Text" />
+<TruncatedTextExample title="Item Display: Truncated Text" />
 
 ## Item Display Props Table
 <ItemDisplayPropsTable />
